Rollback on sync relation
There is a warning after logging in that warns about a known-to-be-empty
relation. This is because when the current session is loaded, the
gebruiker data is not included. Login is done via ember-simple-auth
and adding gebruiker in the includes will be more trouble than marking
this relation sync again and using await.